move emptyTreeWarning
