+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/categories-tree/UseCategoriesTree.tsx
@@ -10,8 +10,7 @@ import subcategorySvg from "@itwin/itwinui-icons/bis-category-subcategory.svg";
 import definitionContainerSvg from "@itwin/itwinui-icons/bis-definitions-container.svg";
 import { createECSqlQueryExecutor } from "@itwin/presentation-core-interop";
 import { HierarchyFilteringPath, HierarchyNodeIdentifier } from "@itwin/presentation-hierarchies";
-import { EmptyTreeContent } from "../common/components/EmptyTreeContent.js";
-import { WarningFilterUnknown, WarningNoMatches, WarningTooManyFilterMatches } from "../common/components/Warnings.js";
+import { EmptyTreeContent, WarningFilterUnknown, WarningNoMatches, WarningTooManyFilterMatches } from "../common/components/WarningEmptyTree.js";
 import { FilterLimitExceededError } from "../common/TreeErrors.js";
 import { useTelemetryContext } from "../common/UseTelemetryContext.js";
 import { CategoriesTreeDefinition } from "./CategoriesTreeDefinition.js";
