+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/common/components/WarningEmptyTree.css
@@ -11,3 +11,14 @@
   padding-right: var(--iui-size-xl);
   padding-left: var(--iui-size-xl);
 }
+
+.tw-empty-tree-warning-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: var(--iui-size-s);;
+  gap: var(--iui-size-xs);
+  width: 100%;
+  height: 100%;
+}
