+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/common/components/WarningEmptyTree.tsx
@@ -3,14 +3,15 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import "./Warnings.css";
-import { Text } from "@itwin/itwinui-react/bricks";
+import "./WarningEmptyTree.css";
+import { Icon, Text } from "@itwin/itwinui-react/bricks";
 import { TreeWidget } from "../../../../TreeWidget.js";
 
-export interface WarningProps {
+interface WarningProps {
   base: string;
 }
 
+/** @internal */
 export function WarningTooManyFilterMatches({ base }: WarningProps) {
   return (
     <div className={"tw-filter-warning-container"}>
@@ -20,6 +21,7 @@ export function WarningTooManyFilterMatches({ base }: WarningProps) {
   );
 }
 
+/** @internal */
 export function WarningNoMatches({ base }: WarningProps) {
   return (
     <div className={"tw-filter-warning-container"}>
@@ -29,10 +31,27 @@ export function WarningNoMatches({ base }: WarningProps) {
   );
 }
 
+/** @internal */
 export function WarningFilterUnknown({ base }: WarningProps) {
   return (
     <div className={"tw-filter-warning-container"}>
       <Text variant={"body-sm"}>{TreeWidget.translate(`${base}.filtering.unknownFilterError`)}</Text>
+    </div>
+  );
+}
+
+interface EmptyTreeContentProps {
+  icon?: string;
+}
+
+/** @internal */
+export function EmptyTreeContent({ icon }: EmptyTreeContentProps) {
+  return (
+    <div className={"tw-empty-tree-warning-container"} style={{}}>
+      {icon ? <Icon size="large" href={icon} /> : null}
+      <Text variant={"body-sm"} style={{ textAlign: "center" }}>
+        {TreeWidget.translate("baseTree.dataIsNotAvailable")}
+      </Text>
     </div>
   );
 }
