+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/models-tree/UseModelsTree.tsx
@@ -13,8 +13,7 @@ import { Anchor, Text } from "@itwin/itwinui-react/bricks";
 import { createECSqlQueryExecutor } from "@itwin/presentation-core-interop";
 import { HierarchyNodeIdentifier, HierarchyNodeKey } from "@itwin/presentation-hierarchies";
 import { TreeWidget } from "../../../TreeWidget.js";
-import { EmptyTreeContent } from "../common/components/EmptyTreeContent.js";
-import { WarningFilterUnknown, WarningNoMatches, WarningTooManyFilterMatches } from "../common/components/Warnings.js";
+import { EmptyTreeContent, WarningFilterUnknown, WarningNoMatches, WarningTooManyFilterMatches } from "../common/components/WarningEmptyTree.js";
 import { useFocusedInstancesContext } from "../common/FocusedInstancesContext.js";
 import { FilterLimitExceededError } from "../common/TreeErrors.js";
 import { useIModelChangeListener } from "../common/UseIModelChangeListener.js";
